Implementation for Sockets
This is work done to lay out how I would like sockets to work in
the system.  I have decided to have `EOD.Socket` be a Protocol
which will have different implementations that can be swapped out
in the future.  The most immediate need for the swap will be test
sockets to use for mocking communication with the client.

I have also elected to add encoding in as part of the socket; my
hope there is to keep a domain specfic message format in the program.
This will come in handy if, in the future, the server supports more
then just one client.