Initial Login System
This is a first pass at the start of a login system to work with
the Dawn of Light Portal login manager.  Currently it will get
a client authenticated with a correct username and password; or, it
will create an account provided the validation passes.

This also introduces the beginning of a database persistence system;
in which the user's very basic account details are currently saved.