Client System First Pass
This introduces the idea of a client and client manager.  The goal
here is for the client manager to dynamically spin up client connections
as the sever passes them to it.

I moved the login logic into the client itself as this feels like a better
place for it.  While I am not fully thrilled with how the deligation of
packets in this process works; I believe we'll have a better process in
place by which to delegate communication with the client into logical
segments of code in the future.