Client Refactor and Packet Bugfix
This introduces a first round pass at refactoring the client with
a more robust packet handling system.  The idea is to make working
with the packets easier and to remove the amount of "noise" that
was happening in each of the packet handlers as much as possible.
I've also started a concept where each packet handler is aware of
what packets it is responsible for.  Right now I'm still using the
simple case statement in the client packet receive loop; however,
in the future I plan to have a routing system which dispatches to
the correct packet handler.

Also in this commit is the introduction of characters.  It is by
no means complete, but just a foot in the door to being able to
create a character from the game client.  With this work I found
a bug that stemmed from my understanding of `:gen_tcp`.  When a
packet is big, such as the character create packet, you don't get
all of the data in a single receive call.  I now detect when this
happens and attempt to fill the rest of the data for the packet
before decoding it.