Working Pass of Character Crud Functionality
This is a testing and working finish for the state of the game in
which a user can now sign up, select a realm and create / delete
characters from the client character screen.

This has uncovered a bug I had in the assumption of `:gen_tcp`;
where I had hoped that calling `recv` would fetch an entire packet;
however, a fairly large packet for the character selection screen
has shown that big packets can come down in chunks.  This forced
me to write a check system to determine if more of a packet is needed
and to continue pulling from the socket until the packet is complete
to send to the client.

I am not fully happy with the encoding and decoding on the characters
of an account; however, I don't want to get bogged down with this layer
as it can be improved independently from the rest of the system.

Another big leak that I would like to clean up later is the fact that
I am saving raw client fields to the database for attributes such as
`hair_color` and `gender`.  Another plan for the future is to figure
out how to map these to more human friendly values or come up with a
mapping system for these values at the very minimum.