New Packet Integeration
This is the code changes needed to integrate the new packet design
work into the existing working system.  With it brings one major
difference; and that is the entire client packet is sent to the
client instead of just the inside data segment.  It will now be up
to the packet handlers to decide what from the client packet they
need.

This commit also removes the now unneeded encoding modules which
only did half of the job that the new packet structure does.  I'm
glad to see them gone :wave: .