Fix Client Restart Bug
The supervisor used to manage clients was incorrectly using the
default restart strategy of `permanent`.  This was forcing the
clients to be restarted even if they were gracefully shutting
down after their socket had been closed.

Not only was this creating zombie clients that weren't connected
to anything.  It was also causing issues with the connection
manager.  I'm not entirely sure how it works, but once a certain
number of failures to read from closed sockets happens it was
closing the main socket in the connection loop.