First Pass for Regions & Server Robustness
The main theme for this commit is the base work to be able to tell
a client which regions are hosted on what IP addresses.  In this
commit I have wired up the server to be passed to the client so the
client can interact with it.

Along with this the server now has a settings structure and has had
some added optional starting paramaters that make it easier to test
stuff without having to open actual network ports.

Of course the biggest focus of the commit was laying down the layout
of regions, which currently is in three basic parts.

* `EOD.Repo.RegionData`
  The persisted information of each reagion; and is used to start
  a `EOD.Region`

* `EOD.Region`
  Currently just a simple GenServer that hold data details from the
  `EOD.Repo.RegionData`

* `EOD.Region.Manager`
  Responsible for booting up a collection of regions and supervising
  them. This is what is used to lookup a region process in the system.