Game Open Request Work
This is the first leg of work towards loading into the world.  This
adds some extra stats to the character repo model which will be needed
very soon as part of the bootstrapping process.  This commit also adds
a new base concept of a `Player`

The idea is that the `Player` will be the nerve center for delegating
in game messages to the client and other server processes.  This is
still a rough sketch and is highly subject to change.