Refactoring and Continue of World Loading
This introduces a big refactor that has been bugging me and I think
needs to take place with the `EOD.Player` genserver.  Initially it
was designed to send some packets as a side effect of starting.  I
feel this is a bad coupling and hides functionality which shouldn't
be tucked away.

The other part of this commit is of course the work done to get some
more packets in for the loading of a character into the world.  Some
of the information being sent is **hard coded** for now but has been
marked as such.  This was done to try and get the client loading, which
is not yet happening but I would like to avoid large swaths of churn in
a single commit.