Introduce Accounts to the SessionManager
It's clear that there will need to be functionality to list all
clients and access them by something meaningful.  I have elected
to use the account usernames for this.  Because there should only
ever be one client per account at any one time this code will also
prevent the account from being used twice to guarantee  uniqueness.

This uniqueness is enforced at login with the client and additional
attempts to login with the same account will correctly notify the
user that the account is already logged in.