Cleaning Up Backend Namespaces
Orginally this project was created with a base namespace of
`EOD.Web` because it was inteaded to be just a web service
backend.  The issue is the Phoenix generator doubled this up
and added extra, non-needed layers and tacked on `Web` to the
end of the given namespace; resulting in `EOD.WebWeb`.

This cleans up the orginazation of the structure; flattening
it and removes all the doubled up namespacing.