Client Admin Display Authenticated Connections
This drills out a bit more of the admin information on the running
game server for corrently authenticated connections.  Currently
it lists all authenticated accounts that are connected; however,
in the future the goal would be to have a filtered view into this
data in case the number get's to be too large.