First Pass for a Working Admin UI
This is a rather big bulk of work that introduces a simple auth
system that currently allows anyone with a game account to log into
the web admin.  The web-admin currently only displays a dashboard
with overview connection information and a more detailed clients
view that displays in real time the open client connections.

With this work in place as a spring-board; more diagnostic and
interactive functionality can be added with a bit more ease to
follow some existing patterns.