Add Client Detail Information and Debugging
This introduces some tooling which allows an admin / dev to get a
list of all connected clients and drill into some of the basic state
of a connected client.

It also allows for some simple packet inspection.  I wanted to do
more with the packets; however, I don't want to muddy the working
base I already have, and the logs currently provide some working
value.

With this work moving to 1.125 should prove to be a bit easier.