Update To Support 1.125d
This gets the code base up to the most recent patch level of live
with it's current set of work.  The need for this is ever increasing
as the plan is to be able to support more devs working on the project
by mid July.

With that in mind; it's not really scalable to try and distribute
older DLLs of the project, nor can they be publicly hosted.  The
goal here is to try and keep the project working with a fresh install
of live so all devs need is a connecting tool to tinker with changes.

Another slight issue that has cropped up is the mix formatter seems
to have diverged between 1.7 and 1.8 of Elixir.  A clean format was
failing on Travis as the 1.7 formatter was finding problems.  Just to
cut this problem off and any other that may come up I've bumped the
Elixir version to 1.8.0.