Update All Remote Dependencies
This is an update of all project dependencies to ensure that the
project doesn't get grossly outdated.  Some of the resulting file
churn is a result from updating credo; which has some updated guides
for how files should be organized.

The other major change is from `comeonin`; which has changed it's
interface with `pbkdf2_elixir`