Prevent Potential Double Query for Region Data
Prior to this; during normal boot of the server the default settings
are initialized.  Part of the default settings is to query for all
enabled regions, which is what also the default option for the region
manager as well.  Without this being wrapped in a `put_new_lazy` the
default was being forced to always query as well.