Correct Controlling Process Flow for Clients
Sockets opened with `:gen_tcp` and the like are are bound to the
process that accepted them as the controlling process.  What does
this mean?  Well, if for some reason the process that accepted
the socket initialy goes down then all of the clients sockets
will close as well.

To avoid this problem from happening you can transfer control of
a socket from one process to another; however, the call must be
made from the process that orginally accepted the connection.

What this fix does is add an additional callback mechanic for
the connection manager to simply call a module/fun with some
optional default arguments.  With this change the server now
boots up the conn manager with a callback of the client manager
directly.  This function now is set to call the needed control
transfer to the started up client; so that if the loop running
crashes not all the clients will close.

This change also has a nice side effect of never needing to go
through the server to get a new client started as it can just
call the client directly.