Refactor EOD.Server to be a Supervisor
This refactor cleans up some of the application supervision and
further removes direct interaction responsibility of the "server"
by changing it from a GenServer to a Supervisor.

Now in stead of a GenServer that starts up other workers, which
felt a bit off to me to begin with, it is now a supervisor that
starts the same workers and gives them names that are registered
under the server name provided or pid.

This work required further work to some of the other workers to
ensure they can be started with names - but it wasn't as bad as
I thought it would be.

This also removes the `ref` option for the server and you can now
only "disable" the connection manager from starting; which is all
I was doing in the tests anyway.