Refactor and Clean up Packet Routing
The packet routing in the client needed cleaned up really bad
before the onslaught of packets that need to be built up with
the subsequent packet handlers.  This also addresses a tiny
bug with the server actually starting in production mode; it
turns out there wasn't a test for that and I missed it after
my last round of refactoring.