Rough Start to Player Data Code Management
The amount of information revolving around a player in the game is
staggering and has a high chance to "run away" in module design.
Looking at Dawn of Light and other similiar open source systems
there always seems to be a rather large `GamePlayer` class that
has a crazy amount of data and logic in it.

The overall goal with this going forward is to attempt to define
different data groups into their own modules and compose a player
as having those different data groups which it references from
a structure.  This is just a first pass that get's the tests going
again; but is by no means a fully solidified pattern / idea yet.