Fix Incorrect Packet Parsing Bug
Prior to this change the code that was checking to see if a packet
was underfilled and then filling up the remainder was blowing up
in the event the client actually had more then just one packet in
the buffer.

Now instead of grabbing everything and parsing; for now I'm electing
to grab the first two bytes; for the length and then pulling exactly
everything I need to fill exactly one packet.  This may have a slight
performance hit since there will always be two recv's for every packet.

If we wanted to pull as much as possible and keep a running excess
the current implementation needs to change quite a bit and I'd rather
research and tackle that later.