Ensure Client Processing Starts at a Known Time
Prior to this the client would beging processing packets coming in
which left a small time frame where the client could crash but it
wasn't set as the controlling process.  This code ensures that a
client will wait to begin it's life-cycle of processing until after
a certain time.  In this case; after it has been tagged as the
controlling process of the socket.